fixed window error, reorganized stuff
GlobalHeader just determines whether to render animated or basic global header based on animated prop.

BasicGlobalHeader is what GlobalHeader was previously.
