+++ packages/gamut-labs/src/experimental/GlobalHeader/AnimatedGlobalHeader/styles.module.scss
@@ -11,7 +11,7 @@
   background-color: white;
   position: fixed;
   top: 0;
-  z-index: 1;
+  z-index: 2;
   width: 100%;
 }
