+++ packages/gamut-labs/src/experimental/GlobalHeader/types.tsx
@@ -11,6 +11,7 @@ type RenderSearch = { desktop: () => ReactNode; mobile: () => ReactNode };
 
 type BaseHeader = {
   action: AppHeaderClickHandler;
+  animated?: boolean;
   className?: string;
   renderSearch?: RenderSearch;
 };
